Move the resolution code into its own package and spread the code out a bit
