+++ sqldelight-compiler/src/main/kotlin/com/squareup/sqldelight/validation/CreateTriggerValidator.kt
@@ -16,8 +16,9 @@
 package com.squareup.sqldelight.validation
 
 import com.squareup.sqldelight.SqliteParser
-import com.squareup.sqldelight.types.ResolutionError
-import com.squareup.sqldelight.types.Resolver
+import com.squareup.sqldelight.resolution.ResolutionError
+import com.squareup.sqldelight.resolution.Resolver
+import com.squareup.sqldelight.resolution.resolve
 import com.squareup.sqldelight.types.Value
 import java.util.ArrayList
