+++ sqldelight-compiler/src/main/kotlin/com/squareup/sqldelight/validation/ExpressionValidator.kt
@@ -16,9 +16,10 @@
 package com.squareup.sqldelight.validation
 
 import com.squareup.sqldelight.SqliteParser
-import com.squareup.sqldelight.types.ResolutionError
-import com.squareup.sqldelight.types.ResolutionError.ExpressionError
-import com.squareup.sqldelight.types.Resolver
+import com.squareup.sqldelight.resolution.Resolver
+import com.squareup.sqldelight.resolution.resolve
+import com.squareup.sqldelight.resolution.ResolutionError
+import com.squareup.sqldelight.resolution.ResolutionError.ExpressionError
 import com.squareup.sqldelight.types.Value
 import org.antlr.v4.runtime.RuleContext
